remove WithWBRs
The site is currently working locally and broken in production. Hard to
troubleshoot. This is one attempt to start troubleshooting.

On the production site, it says that the problem is in the `WithWBRs`
file. But.

The way Astro builds confuses me. In this project, the `dist/_astro`
directory ends up with a bunch of mostly-small JS files, and one
enormous `WithWBRs.js` file (1.8Mb!). If I remove this file, then it
ends up with a equally-gigantic `NearLogin.js`.

Ok, so for some reason Astro really likes the name of the last
alphabetical file in the `components/` subfolder. Why?

And does the production problem have anything to do with the actual
WithWBRs? Probably not. Let's see if removing it clarifies anything.